reduce work with deleting resume
