+++ app/config/routing.yml
@@ -26,17 +26,6 @@ app_movies:
         resource: "@AppBundle/Resources/config/routing/movies.yml"
         prefix:   /movies
 
-##############################################
-#                                            #
-#                  RESUME                    #
-#                                            #
-##############################################
-
-app_resume:
-    resource: "@AppBundle/Resources/config/routing/resume.yml"
-    prefix:   /resume
-
-
 ############################################################################################################
