+++ src/AppBundle/Controller/MoviesController.php
@@ -4,7 +4,6 @@ namespace AppBundle\Controller;
 
 use AppBundle\Entity\Movies;
 use AppBundle\Entity\Category;
-use AppBundle\Entity\Resume;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -54,9 +53,7 @@ class MoviesController extends Controller
     public function newAction(Request $request)
     {
         $movie = new Movies();
-        $resume = new Resume();
         $form = $this->createForm('AppBundle\Form\MoviesType', $movie);
-        $form = $this->createForm('AppBundle\Form\ResumeType', $resume);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -84,7 +81,6 @@ class MoviesController extends Controller
 
         return $this->render('movies/new.html.twig', array(
             'movie' => $movie,
-            'resume' => $resume,
             'form' => $form->createView(),
         ));
     }
