--- src/AppBundle/Resources/config/doctrine/Resume.orm.yml
@@ -1,28 +1,0 @@
-AppBundle\Entity\Resume:
-    type: entity
-    table: null
-    repositoryClass: AppBundle\Repository\ResumeRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        resume:
-            type: text
-            nullable: true
-        author:
-            type: string
-            length: 255
-            nullable: true
-        actors:
-            type: string
-            length: 255
-            nullable: true
-
-    oneToOne:
-        movies:
-          targetEntity: AppBundle\Entity\Movies
-          cascade: ['all']
-    lifecycleCallbacks: {  }
